Main em server class easier to reuse
Untangles the main server class of the event manager from command line
options, daemonizing code and pid-files. This allows the server to be
included in other software without the need to stub and mock these
parts.
